Move MatrixVectorIterator and NonzeroIterator into utils/iterators instead of utils/
