+++ include/graphblas/reference/blas3.hpp
@@ -25,7 +25,7 @@
 #include <type_traits> //for std::enable_if
 
 #include <graphblas/base/blas3.hpp>
-#include <graphblas/utils/MatrixVectorIterator.hpp>
+#include <graphblas/utils/iterators/MatrixVectorIterator.hpp>
 
 #include "io.hpp"
 #include "matrix.hpp"
