+++ include/graphblas/reference/matrix.hpp
@@ -49,7 +49,6 @@
 #include <graphblas/type_traits.hpp>
 
 #include <graphblas/algorithms/hpcg/ndim_matrix_builders.hpp>
-#include <graphblas/utils/MatrixVectorIterator.hpp>
 #include <graphblas/utils/iterators/utils.hpp>
 
 #include "non_zero_wrapper.h"
