+++ tests/unit/clearMatrix.cpp
@@ -18,7 +18,7 @@
 #include <iostream>
 #include <sstream>
 
-#include <graphblas/utils/MatrixVectorIterator.hpp>
+#include <graphblas/utils/iterators/MatrixVectorIterator.hpp>
 
 #include <graphblas.hpp>
